drawsync-backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pymupdf>=1.24.0 ; python_version < "3.13"
pillow>=10.0
openai>=1.30.0
python-dotenv>=1.0.1
email-validator>=2.1
google-cloud-vision>=3.7.2
PyJWT>=2.8.0
httpx>=0.27.0
resend>=2.12.0